Commit #4: Added the use of items, still need to allow the use of their functionality through a menu. Modified the pause menu, can now use back command by pressing 'B' on the 360 controller. Entities can now be added and removed when needed in the pause menu.

+++ CollisionboxUpdateSystem.h
@@ -48,7 +48,7 @@ public:
 	{
 		PositionComponent& p = *positionMapper.get(e);
 		shape.setPosition(p.getPosX() - width / 2, p.getPosY() - height / 2);
-		shape.setSize(sf::Vector2f(collisionBox->width, collisionBox->height));
+		shape.setSize(sf::Vector2f((float)collisionBox->width, (float)collisionBox->height));
 		shape.setOutlineThickness(2.0f);
 		shape.setOutlineColor(sf::Color::Red);
 		shape.setFillColor(sf::Color::Transparent);
@@ -56,7 +56,7 @@ public:
 	}
 
 private:
-	float width, height;
+	int width, height;
 	sf::Sprite* sprite;
 	sf::RectangleShape shape;
 	sf::RenderWindow* window;
